Initial Angular setup (#50)
* Initial Angular setup

Sets up a barebones Angular project.

* Set up client for Angular

Sets up the A2A client and processor for the Angular example.

* Expose more types from web

Exposes the primitives and `Action` for the Angular implementation.

* Update Angular client

Updates the client for Angular to account for recent changes.

* Shared logic for Angular rendering

Sets up some logic that will be shared across components in the Angular implementation.

* Add initial implementations of Angular components

Adds initial implementations for the various components in the catalog in Angular.

* Account for processor changes

Updates the Angular components to account for recent changes in the shared processor.

* Update Angular readme

Updates the readme for the Angular app.

* Add license to Angular files

Updates all Angular files to add a license header.