[angular] Rework rendering (#58)
* [angular] Account for spec change

Accounts for the changes to the button spec.

* [angular] Rework rendering

Reworks the Angular renderer to be based on a config object, rather than rendering components through the template. This reduces the number of host nodes which should make it easier to style the components. It also allows users to provide their own component catalog.