Update Angular setup (#236)
* [angular] Add usageHint to Text

Updates the `Text` component to account for `usageHint`.

* [angular] Remove Heading from default catalog

The `Heading` was replaced by `Text` plus a `usageHint` so we can remove it.

* [angular] Add usageHint to Image

Updates the `Image` component to handle `usageHint`.

* Fix typo in Lit component

The `Text` component in Lit was using the styles for `Image`.

* [angular] Update to Angular v21

Updates the repo to the latest version of Angular.

* [angular] Update themes

Syncs up the Angular demo themes with the Lit ones.